restructure input output module
