+++ src/gating/contour_based_gating.py
@@ -12,7 +12,7 @@ from gating.automatic_gating import AutomaticGating
 from gui.popup_windows.message_boxes import ErrorMessage
 from gui.popup_windows.frame_range_dialog import FrameRangeDialog, StartFramesDialog
 from gui.right_half.right_half import toggle_diastolic_frame, toggle_systolic_frame
-from report.report import report
+from input_output.output.reports import report
